requirements.txt
aws-cdk-lib==2.214.0
constructs==10.4.2